requirements.txt
Pillow==8.3.1
pygubu==0.12
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
